added new features
I successfully implemented and edit and delete function to the sales list. The edit page was created and the ID of the sale and all corresponding data was added to that had a page so information is preloaded when edit is required. Search sales was added. Allowing the user to enter and search criteria and find items based on search terms. Styles were implemented to maintain uniform website design.